Refactored popover
