+++ package.json
@@ -13,7 +13,7 @@
   "license": "ISC",
   "main": "index.js",
   "name": "jsonschema-form-for-material-ui",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "repository": {
     "type": "git",
     "url": "https://github.com/TwoAbove/jsonschema-form-for-material-ui.git"
