+++ src/fields/ConfiguredField.jsx
@@ -6,43 +6,12 @@ import FormControl from '@material-ui/core/FormControl';
 import FormHelperText from '@material-ui/core/FormHelperText';
 import Input from '@material-ui/core/Input';
 import Typography from '@material-ui/core/Typography';
-import Popover from '@material-ui/core/Popover';
-import IconButton from '@material-ui/core/IconButton';
-import InfoIcon from '@material-ui/icons/InfoOutlined';
-import debounce from 'lodash.debounce';
 
 import fieldStyles from './field-styles';
-
-const getLTofEl = el => {
-  const box = el.getBoundingClientRect();
-  return {
-    top: box.top,
-    left: box.left,
-  };
-}
+import PopoverInfo from './components/PopoverInfo';
 
 // for unit testing only
 export class RawConfiguredField extends React.Component {
-  state = {
-    anchorEl: null,
-  };
-
-  getAnchorPosition = () => {
-    if (!this.state.anchorEl) return { top: 0, left: 0 };
-    const pos = getLTofEl(this.state.anchorEl);
-    if (this.props.popUpOffset) {
-      if (this.props.popUpOffset.top) pos.top += this.props.popUpOffset.top;
-      if (this.props.popUpOffset.left) pos.left += this.props.popUpOffset.left;
-    }
-    return pos;
-  }
-
-  handlePopoverOpen = debounce(event => {
-    this.setState({ anchorEl: event.currentTarget })
-  }, 250, { leading: true });
-  handlePopoverClose = debounce(() => {
-    this.setState({ anchorEl: null })
-  }, 250, { leading: true });
 
   shouldComponentUpdate = (nextProps, nextState) => {
     if (this.props.data !== nextProps.data) return true;
@@ -66,8 +35,7 @@ export class RawConfiguredField extends React.Component {
         componentProps = {},
         id,
     } = this.props;
-    const { anchorEl } = this.state;
-    const open = Boolean(anchorEl);
+
     return (
       <FormControl className={classNames(classes.root, { [classes.withLabel]: LabelComponent })}>
         {LabelComponent && title &&
@@ -75,37 +43,11 @@ export class RawConfiguredField extends React.Component {
               {title}
           </LabelComponent>)
         }
-
-        {descriptionText ? <IconButton
-          style={{position: 'absolute', left: -20, top: 15, width: 20, height: 20}}
-          aria-owns={open ? 'mouse-over-popover' : null}
-          aria-haspopup="true"
-          disableRipple
-          className={classes.infoButton}
-          disableTouchRipple
-          onMouseEnter={this.handlePopoverOpen}
-          onMouseLeave={this.handlePopoverClose}
-          >
-          <InfoIcon />
-        </IconButton>
-        : null}
-        {descriptionText ? <Popover
-            anchorReference="anchorPosition"
-            id="mouse-over-popover"
-            className={classes.infoPopover}
-            classes={{
-              paper: classes.infoPaper,
-            }}
-            style={{ pointerEvents: 'none' }}
-            open={open}
-            anchorEl={anchorEl}
-            anchorPosition={this.getAnchorPosition()}
-            onClose={this.handlePopoverClose}
-            disableRestoreFocus
-          >
-            <Typography>{descriptionText}</Typography>
-          </Popover> : null}
-
+        {descriptionText ?
+          <PopoverInfo
+            descriptionText={descriptionText}
+            classes={classes}
+          /> : null}
         <Component
           className={className && classes[className]}
           value={data}
